update user controller

+++ clouddo-admin/src/main/java/com/bootdo/clouddoadmin/service/impl/UserServiceImpl.java
@@ -1,17 +1,23 @@
 package com.bootdo.clouddoadmin.service.impl;
 
 
-import com.bootdo.clouddoadmin.domain.Tree;
 
-import com.bootdo.clouddoadmin.utils.*;
+
+import com.alibaba.fastjson.JSONObject;
 import com.bootdo.clouddoadmin.dao.DeptDao;
 import com.bootdo.clouddoadmin.dao.UserDao;
 import com.bootdo.clouddoadmin.dao.UserRoleDao;
 import com.bootdo.clouddoadmin.domain.DeptDO;
+import com.bootdo.clouddoadmin.domain.Tree;
 import com.bootdo.clouddoadmin.domain.UserDO;
 import com.bootdo.clouddoadmin.domain.UserRoleDO;
+import com.bootdo.clouddoadmin.enums.ErrorCode;
 import com.bootdo.clouddoadmin.service.UserService;
+import com.bootdo.clouddoadmin.utils.BuildTree;
+import com.bootdo.clouddoadmin.utils.MD5Utils;
 import com.bootdo.clouddoadmin.vo.UserVO;
+import com.bootdo.clouddocommon.utils.Response;
+import com.bootdo.clouddocommon.utils.Responses;
 import org.apache.commons.lang.ArrayUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -19,7 +25,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
-
 
 import java.util.*;
 
@@ -58,8 +63,22 @@ public class UserServiceImpl implements UserService {
 	}
 
 	@Override
-	public int save(UserDO user) {
+	public Response<Integer> save(UserDO user) {
+		//验证用户是否已存在
+		Response<UserDO> existUserResponse =  queryUserByMobile(user.getMobile());
+		if(!existUserResponse.isSuccess()) {
+			logger.error("query user by mobile failed when save user , mobile: {}", user.getMobile());
+			return Responses.fail(ErrorCode.USER_QUERY_ERROR.getCode(), ErrorCode.USER_QUERY_ERROR.getMsg());
+		}
+		if(existUserResponse.getData() != null) {
+			logger.error("user has exist, mobile : {}", user.getMobile());
+			return Responses.fail(ErrorCode.USER_EXIST_ERROR.getCode(), ErrorCode.USER_EXIST_ERROR.getMsg());
+		}
 		int count = userMapper.save(user);
+		if(count == 0) {
+			logger.error("save user failed, user : {}", JSONObject.toJSON(user));
+			return Responses.fail(ErrorCode.USER_SAVE_ERROR.getCode(), ErrorCode.USER_SAVE_ERROR.getMsg());
+		}
 		Long userId = user.getUserId();
 		List<Long> roles = user.getroleIds();
 		userRoleMapper.removeByUserId(userId);
@@ -73,12 +92,16 @@ public class UserServiceImpl implements UserService {
 		if (list.size() > 0) {
 			userRoleMapper.batchSave(list);
 		}
-		return count;
+		return Responses.of(count);
 	}
 
 	@Override
-	public int update(UserDO user) {
+	public Response<Integer> update(UserDO user) {
 		int r = userMapper.update(user);
+		if(r == 0) {
+			logger.error("update user failed, user : {}", JSONObject.toJSON(user));
+			return Responses.fail(ErrorCode.USER_SAVE_ERROR.getCode(), ErrorCode.USER_SAVE_ERROR.getMsg());
+		}
 		Long userId = user.getUserId();
 		List<Long> roles = user.getroleIds();
 		if(null!=roles){
@@ -94,13 +117,14 @@ public class UserServiceImpl implements UserService {
 				userRoleMapper.batchSave(list);
 			}
 		}
-		return r;
+		return Responses.of(r);
 	}
 
 	@Override
-	public int remove(Long userId) {
+	public Response<Integer> remove(Long userId) {
 		userRoleMapper.removeByUserId(userId);
-		return userMapper.remove(userId);
+		int delNum = userMapper.remove(userId);
+		return Responses.of(delNum);
 	}
 
 	@Override
@@ -115,7 +139,7 @@ public class UserServiceImpl implements UserService {
 	}
 
 	@Override
-	public int resetPwd(UserVO userVO,UserDO userDO) throws Exception {
+	public int resetPwd(UserVO userVO, UserDO userDO) throws Exception {
 		if(Objects.equals(userVO.getUserDO().getUserId(),userDO.getUserId())){
 			if(Objects.equals(MD5Utils.encrypt(userDO.getUsername(),userVO.getPwdOld()),userDO.getPassword())){
 				userDO.setPassword(MD5Utils.encrypt(userDO.getUsername(),userVO.getPwdNew()));
@@ -193,6 +217,12 @@ public class UserServiceImpl implements UserService {
 	@Override
 	public Map<String, Object> updatePersonalImg(MultipartFile file, String avatar_data, Long userId) throws Exception {
 		return null;
+	}
+
+	@Override
+	public Response<UserDO> queryUserByMobile(String mobile) {
+		UserDO userDO = userMapper.getUserByMobile(mobile);
+		return Responses.of(userDO);
 	}
